requirements.txt
streamlit>=1.29.0
python-dotenv>=0.19.0
pinecone-client>=2.2.4
langchain>=0.0.350
langchain-openai>=0.0.2
openai>=1.3.0
numpy>=1.21.0
pandas>=1.3.0
pydantic>=2.0.0
python-dateutil>=2.8.2
tqdm>=4.65.0
